fix(a7670c_tls.hpp): httpsPost возвращает int (HTTP-код или отрицательный код ошибки)
- было: uint16_t — не покрывал отрицательные коды ошибок
- стало: int — соответствует реализации в .cpp и вызову в app.cpp
Ref: ocean-station/gsm/cipopen_transport

+++ Core/Inc/a7670c_tls.hpp
@@ -32,7 +32,7 @@ public:
     int  connect(const char* host, uint16_t port);
 
     // HTTPS POST (вызывать после connect)
-    uint16_t httpsPost(const char* url, const char* json, uint16_t len);
+    int      httpsPost(const char* url, const char* json, uint16_t len);
 
     // Закрыть соединение
     void close();
